Add 'generations' concept
I thought it would be great if the caller
had access to all individual iterations
of an L-system. This prompted me to
change the implementation of the
algorithm to provide a lazy list
of iterations. Here, I decided to name
the concept of an L-system iteration
'generation'.
Retrieving the specific iteration that
you're interested in is still possible
with the 'generation' fn, which takes
an extra integer parameter denominating
the number of the generation that you
are interested in.