Change sprout design - must return vector
When a function is used as a value in a rules
map instead of a vector, linden/generations
would replace the rule key with the return
value of that function. However, this meant
that it would be impossible for functions
to return multiple values. So, instead,
the design has been changed so that
functions used as values in a rules map
must return a vector, regardless of the
desired number of return values. Using
linden/generations, the key will then
be replaced with the contents of the vector,
just as when you use a vector as a value.

Apart from changing the replacement logic,
this commit also fixes the unit tests
that are affected by the logic change.