Replace mundane recursion with loop/recur
because in a cljs project using Linden,
I experienced "too much recursion" error
messages in Firefox when calling
l/generations with a certain ruleset
and calling (take % 5) on the returned
lazy seq. I can only assume that this is
because the call stack is blown due to
mundane recursion.

Remove the function called
"nth-generation" because it is unsuitable
when random elements in the ruleset cause
non-idempotent lazy seqs to be returned
from l/generations.

Fix tests so that they now call
l/generations instead of l/nth-generation.

Remove redundant comments.